% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wget.R
\name{wget_set}
\alias{wget_set}
\title{wget_set}
\usage{
wget_set(method = "wget", extra = "-c")
}
\arguments{
\item{method}{download method (wget, default)}

\item{extra}{extra parameter (-c, default)}
}
\description{
set "download.file.method" to "wget" and "download.file.extra" to "-c"
}
\details{
setting download method to "wget -c", so that it can continue the download from where it left off.
}
\examples{
wget_set()
}
\author{
Guangchuang Yu
}
