% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wget.R
\name{wget_unset}
\alias{wget_unset}
\title{wget_unset}
\usage{
wget_unset()
}
\description{
unset download method setting by \code{wget_set}
}
\details{
set the download method and extra parameter back to what it is or NULL if it is not specify
}
\author{
Guangchuang Yu
}
